requirements.txt
functions-framework==3.*
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.126.0
google-cloud-secret-manager==2.20.0
google-cloud-storage==2.16.0
google-cloud-bigquery==3.22.0
python-dotenv==1.0.1
simpledbf==0.2.6
tenacity==8.3.0
aiofiles==23.2.1
pyzipper==0.3.6
